reorganize, need to add the perks to the handlers
